refactor: Replace flat diagnostic system with inheritance-based architecture
Major refactoring of the diagnostic system to improve maintainability and extensibility:

## Changes
- Created IPSDiagnostics and DRDiagnostics dataclasses with proper inheritance
- DRDiagnostics extends IPSDiagnostics since DR uses IPS internally
- Updated all estimators to produce diagnostic objects instead of dicts
- Fixed visualization to work with both new objects and legacy dicts
- Added proper type hints and fixed all mypy errors
- Fixed DR diagnostics display in analyze_dataset.py

## Benefits
- Clear separation between IPS and DR diagnostics
- Type-safe diagnostic objects with validation
- Reduced code duplication in DR estimators
- Better extensibility for future diagnostic types
- Backward compatibility with existing visualization code

## Testing
- All existing tests pass
- New tests added for diagnostic objects
- Mypy type checking passes

+++ cje/utils/diagnostics/dr.py
@@ -14,7 +14,7 @@ logger = logging.getLogger(__name__)
 
 def _p_value_from_z(z: float) -> float:
     """Convert z-score to two-sided p-value."""
-    return 2 * (1 - stats.norm.cdf(abs(z)))
+    return float(2 * (1 - stats.norm.cdf(abs(z))))
 
 
 def compute_dr_policy_diagnostics(
@@ -28,7 +28,7 @@ def compute_dr_policy_diagnostics(
     policy: str = "unknown",
 ) -> Dict[str, Any]:
     """Compute comprehensive DR diagnostics for a single policy.
-    
+
     Args:
         dm_component: Direct method (outcome model) component
         ips_correction: IPS correction component
@@ -38,12 +38,12 @@ def compute_dr_policy_diagnostics(
         influence_functions: Per-sample influence functions
         unique_folds: Unique fold IDs used in cross-fitting
         policy: Policy name
-    
+
     Returns:
         Dictionary with diagnostic metrics
     """
     n = len(dm_component)
-    
+
     diagnostics = {
         "policy": policy,
         "n_samples": n,
@@ -53,7 +53,7 @@ def compute_dr_policy_diagnostics(
         "dm_std": float(dm_component.std()),
         "ips_corr_std": float(ips_correction.std()),
     }
-    
+
     # Outcome model fit (if fresh rewards available)
     if fresh_rewards is not None and outcome_predictions is not None:
         mask = ~np.isnan(fresh_rewards) & ~np.isnan(outcome_predictions)
@@ -62,22 +62,22 @@ def compute_dr_policy_diagnostics(
             diagnostics["residual_mean"] = float(residuals.mean())
             diagnostics["residual_std"] = float(residuals.std())
             diagnostics["residual_rmse"] = float(np.sqrt((residuals**2).mean()))
-            
+
             # R² (out-of-fold if cross-fitted)
             ss_res = np.sum(residuals**2)
-            ss_tot = np.sum((fresh_rewards[mask] - fresh_rewards[mask].mean())**2)
+            ss_tot = np.sum((fresh_rewards[mask] - fresh_rewards[mask].mean()) ** 2)
             r2 = 1 - ss_res / max(ss_tot, 1e-12)
             diagnostics["r2_oof"] = float(r2)
         else:
             diagnostics["r2_oof"] = np.nan
             diagnostics["residual_rmse"] = np.nan
-    
+
     # Influence function diagnostics
     if influence_functions is not None:
         diagnostics["if_mean"] = float(influence_functions.mean())
         diagnostics["if_std"] = float(influence_functions.std())
         diagnostics["if_var"] = float(influence_functions.var())
-        
+
         # Check for heavy tails
         diagnostics["if_tail_ratio_99_5"] = tail_weight_ratio(
             np.abs(influence_functions), 0.05, 0.99
@@ -85,7 +85,7 @@ def compute_dr_policy_diagnostics(
         diagnostics["if_top1_mass"] = mass_concentration(
             np.abs(influence_functions), 0.01
         )
-        
+
         # Score function test (should be mean zero for TMLE)
         score_mean = influence_functions.mean()
         score_se = influence_functions.std() / np.sqrt(n)
@@ -94,7 +94,7 @@ def compute_dr_policy_diagnostics(
         diagnostics["score_se"] = float(score_se)
         diagnostics["score_z"] = float(score_z)
         diagnostics["score_p"] = _p_value_from_z(score_z)
-    
+
     # Cross-fitting info
     if unique_folds is not None:
         diagnostics["cross_fitted"] = True
@@ -102,7 +102,7 @@ def compute_dr_policy_diagnostics(
     else:
         diagnostics["cross_fitted"] = False
         diagnostics["unique_folds"] = 1
-    
+
     # Coverage check (do we have enough fresh draws?)
     diagnostics["coverage_ok"] = True
     if fresh_rewards is not None:
@@ -111,11 +111,11 @@ def compute_dr_policy_diagnostics(
         if coverage < 0.8:
             diagnostics["coverage_ok"] = False
             logger.warning(f"Low fresh draw coverage for {policy}: {coverage:.1%}")
-    
+
     # Component correlation (ideally low for orthogonality)
     corr = np.corrcoef(dm_component, ips_correction)[0, 1]
     diagnostics["component_correlation"] = float(corr) if not np.isnan(corr) else 0.0
-    
+
     return diagnostics
 
 
@@ -124,25 +124,25 @@ def compute_dr_diagnostics_all(
     influence_functions: Optional[Dict[str, np.ndarray]] = None,
 ) -> Dict[str, Any]:
     """Compute DR diagnostics for all policies.
-    
+
     Args:
         estimator: A DR estimator with _dm_component and _ips_correction
         influence_functions: Optional dict of influence functions by policy
-    
+
     Returns:
         Dictionary with per-policy diagnostics and summary metrics
     """
     all_diagnostics = {}
-    
+
     for policy in estimator.sampler.target_policies:
         # Get components
         dm = estimator._dm_component.get(policy)
         ips = estimator._ips_correction.get(policy)
-        
+
         if dm is None or ips is None:
             logger.warning(f"Missing DR components for {policy}")
             continue
-        
+
         # Get optional data
         fresh_rewards = None
         outcome_preds = None
@@ -150,12 +150,12 @@ def compute_dr_diagnostics_all(
             fresh_rewards = estimator._fresh_rewards.get(policy)
         if hasattr(estimator, "_outcome_predictions"):
             outcome_preds = estimator._outcome_predictions.get(policy)
-        
+
         # Get influence functions if available
         ifs = None
         if influence_functions and policy in influence_functions:
             ifs = influence_functions[policy]
-        
+
         # Compute diagnostics
         all_diagnostics[policy] = compute_dr_policy_diagnostics(
             dm_component=dm,
@@ -166,5 +166,5 @@ def compute_dr_diagnostics_all(
             influence_functions=ifs,
             policy=policy,
         )
-    
+
     return all_diagnostics
